+++ Pipfile
@@ -1,0 +1,18 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+pytest = "==7.2.2"
+numpy = "*"
+pandas = "*"
+matplotlib = "*"
+seaborn = "*"
+flask = "==3.0.0"
+geopandas = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.10"
