--- requirements.txt
@@ -1,11 +1,0 @@
-pytest==7.2.2
-
-numpy==1.26.0
-
-pandas==1.3.5
-
-matplotlib==3.6.0
-
-seaborn==0.13.0
-
-flask==3.0.0
